feat(website): move language translations to client side
