+++ website/app/components/preview/Language.tsx
@@ -1,5 +1,4 @@
 import { useSelector } from '@legendapp/state/react';
-import { useFetcher } from '@remix-run/react';
 import { useEffect } from 'react';
 
 import { DropdownSimple } from '@/components/Dropdown';
@@ -8,33 +7,18 @@ import { subsetToLanguage } from '@/utils/language/subsets';
 import { previewState } from './observables';
 
 interface LanguageSelectorProps {
+	defSubset: string;
 	subsets: string[];
 }
 
-interface LanguageFetcher {
-	text: string;
-}
-
-const LanguageSelector = ({ subsets }: LanguageSelectorProps) => {
+const LanguageSelector = ({ defSubset, subsets }: LanguageSelectorProps) => {
 	const language = useSelector(previewState.language);
-	const languageFetcher = useFetcher<LanguageFetcher>();
 
 	useEffect(() => {
-		languageFetcher.submit(
-			{ subset: language },
-			{
-				method: 'POST',
-				action: '/actions/language',
-			},
-		);
-		// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, [language]);
-
-	useEffect(() => {
-		if (languageFetcher.state === 'idle' && languageFetcher.data?.text) {
-			previewState.text.set(languageFetcher.data.text);
+		if (defSubset !== 'latin') {
+			previewState.language.set(defSubset);
 		}
-	}, [languageFetcher]);
+	}, [defSubset]);
 
 	const items = subsets.map((lang) => ({
 		label: subsetToLanguage(lang),
